# Revert last changes that caused performance regressions