@He-Pin 2025-04-29
Wired, The lazy val will access the barrier