@He-Pin 2025-04-29
If this could work, then a var evaluater  and evaluater() method to get or create will works too , and avoid the double check, wired any way.